New Development: No             JIRA Issue: No         Ready for Test: Yes      Interface Changes: No What Interface Changed: Please list interface changes          Test Programs: List test programs   Put in Release Notes: Yes/No              Module(s): Module Names change impacts.            Description: fix for 32 bit compile issues
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk@22805 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ casacore/ms/MeasurementSets/MSSummary.cc
@@ -1787,7 +1787,7 @@ void MSSummary::listSpectralAndPolInfo (LogIO& os, Bool verbose) const
 			vector<String>::const_iterator iter=names.begin();
 			iter!=names.end(); iter++
 		) {
-			widthName = max(widthName, iter->size());
+			widthName = max(widthName, (Int)iter->size());
 		}
 
 		// Define the column widths
